+++ HISTORY.rst
@@ -6,3 +6,10 @@ History
 ------------------
 
 * First release on PyPI.
+
+0.7.0 (2017-07-28)
+------------------
+
+* Implement :code:`VersionedSerializer`
+* Implement :code:`view_config`
+* Make the library an installable Django app